+++ src/file.rs
@@ -1,6 +1,6 @@
 use std::fs::File;
 use std::io::Read;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 
 use crate::errors::*;
 
@@ -18,4 +18,20 @@ fn read_(path: &Path) -> Result<String> {
         .read_to_string(&mut s)
         .wrap_err_with(|| format!("couldn't read {}", path.display()))?;
     Ok(s)
+}
+
+pub fn canonicalize(path: impl AsRef<Path>) -> Result<PathBuf> {
+    _canonicalize(path.as_ref())
+}
+
+fn _canonicalize(path: &Path) -> Result<PathBuf> {
+    #[cfg(target_os = "windows")]
+    {
+        // Docker does not support UNC paths, this will try to not use UNC paths
+        dunce::canonicalize(&path).map_err(Into::into)
+    }
+    #[cfg(not(target_os = "windows"))]
+    {
+        Path::new(&path).canonicalize().map_err(Into::into)
+    }
 }